refactor(shell): notification auth residual 분리
